refactor: rename integrate-kobo-button and extract kobo-configuration-dialog
