+++ interfaces/portal/src/app/pages/program-settings-registration-data/components/kobo-integration-card/kobo-integration-card.component.html
@@ -1,0 +1,47 @@
+<app-card-with-link
+  title="KoboToolbox"
+  i18n-title
+  [titleColoredChipLabel]="titleColoredChipLabel()"
+  [titleColoredChipColor]="'green'"
+  (cardClicked)="
+    isKoboIntegrated() ? undefined : koboConfigurationDialog.show()
+  "
+  [subtitle]="cardSubtitle()"
+  [image]="'assets/registration-data/kobo-toolbox.svg'"
+  [enableLink]="!isKoboIntegrated()"
+>
+  <ng-container card-content>
+    @if (isKoboIntegrated()) {
+      <div class="absolute inset-e-4 top-2">
+        <app-ellipsis-menu [menuItems]="menuItems()" />
+      </div>
+
+      <a
+        [href]="externalFormUrl()"
+        [rel]="'external'"
+        target="_blank"
+        title="Opens in a new window"
+        i18n-title="@@generic-opens-in-new-window"
+        class="inline-block [&_span:first-child]:underline hover:[&_span:first-child]:no-underline focus:[&_span:first-child]:no-underline"
+        ><span>{{ koboIntegration.data()?.name ?? externalFormUrl() }}</span>
+        <span class="p-button-icon pi pi-external-link ms-1 text-sm"></span>
+      </a>
+    }
+  </ng-container>
+  <ng-container card-footer>
+    @if (isKoboIntegrated()) {
+      <p
+        i18n
+        class="text-end"
+      >
+        Last updated:
+        <span>{{ koboIntegration.data()?.dateDeployed | date }}</span>
+      </p>
+    }
+  </ng-container>
+</app-card-with-link>
+
+<app-kobo-configuration-dialog
+  #koboConfigurationDialog
+  [programId]="programId()"
+/>
