+++ interfaces/portal/src/app/pages/program-settings-registration-data/program-settings-registration-data.page.html
@@ -12,9 +12,9 @@
     >
       Select a registration tool you want to integrate to the program
     </p>
-    <app-integrate-kobo-button
+    <app-kobo-integration-card
       [programId]="programId()"
       class="w-1/2"
-    ></app-integrate-kobo-button>
+    ></app-kobo-integration-card>
   </p-card>
 </app-page-layout-program-settings>
