+++ interfaces/portal/src/app/pages/program-settings-registration-data/program-settings-registration-data.page.ts
@@ -3,14 +3,14 @@ import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { CardModule } from 'primeng/card';
 
 import { PageLayoutProgramSettingsComponent } from '~/components/page-layout-program-settings/page-layout-program-settings.component';
-import { IntegrateKoboButtonComponent } from '~/pages/program-settings-registration-data/components/integrate-kobo-button/integrate-kobo-button.component';
+import { KoboIntegrationCardComponent } from '~/pages/program-settings-registration-data/components/kobo-integration-card/kobo-integration-card.component';
 
 @Component({
   selector: 'app-program-settings-registration-data',
   imports: [
     CardModule,
     PageLayoutProgramSettingsComponent,
-    IntegrateKoboButtonComponent,
+    KoboIntegrationCardComponent,
   ],
   templateUrl: './program-settings-registration-data.page.html',
   styles: ``,
